+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- event asset migration (#750)
+
 ## [0.13.2] - 2022-05-24
 
 ### Fixed